requirements.txt
flask~=3.0.0
openai~=0.28.1
langchain~=0.0.325
python-dotenv~=1.0.0